requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5.2
backcall==0.1.0
bleach==3.1.4
cffi==1.14.0
channels==2.1.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
distlib==0.3.0
Django==3.0.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
parso==0.6.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
pyAesCrypt==0.4.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted @ file:///C:/Users/chinm/Downloads/Twisted-20.3.0-cp38-cp38-win_amd64.whl
txaio==20.4.1
virtualenv==20.0.20
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zope.interface==5.1.0
